# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Voltpilot-EMS Portal API
  version: 0.1.0-mvp
  description: >
    OpenAPI contract for the portal backend (Spring Boot). BINDING contract
    (architecture section 20 item 8). Covers the MVP portal surface: auth-guarded
    tenancy, device claiming, telemetry and schedule views, and economic KPIs
    (architecture section 4).


    Every non-system endpoint requires a Keycloak-issued Bearer JWT (realm
    `voltpilot`). The `tenant_id` claim scopes all data via Postgres
    Row-Level-Security - a caller only ever sees its own tenant's rows; another
    tenant's resources return 404 (not found) rather than 403. The
    sites/devices/telemetry/prices/weather/schedule and device-claim endpoints
    are implemented in services/api; KPIs remain a stub pending a later
    increment.
  contact:
    name: Voltpilot-EMS
servers:
  - url: http://localhost:8090
    description: Local dev (docker compose / `mvn spring-boot:run`)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/tenants:
    get:
      tags: [admin]
      summary: List all tenants (platform-wide).
      responses:
        "200":
          description: Every tenant on the platform.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Create a tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTenantRequest" }
      responses:
        "201":
          description: Tenant created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/tenants/{tenantId}:
    put:
      tags: [admin]
      summary: Update a tenant's master data (name/segment).
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTenantRequest" }
      responses:
        "200":
          description: Updated tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "400": { description: Invalid input (blank name or bad segment). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
  /api/v1/admin/tenants/{tenantId}/delete:
    post:
      tags: [admin]
      summary: Offboard a tenant (type-to-confirm full cascade).
      description: >
        The most destructive platform action. The body must carry the tenant's
        EXACT name (`confirmName`); a mismatch is refused (400) before anything
        is touched. The database cascade - all series data, sites, devices,
        assets and the tenant row - runs in ONE transaction. The tenant's
        retained MQTT topics and Keycloak users are then cleaned best-effort;
        users whose deletion failed are reported by name in `failedUsers` so a
        partial directory failure is never silent.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteTenantRequest" }
      responses:
        "200":
          description: Tenant offboarded; the report of what was removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantOffboardingReport" }
        "400": { description: confirmName does not match the tenant's name. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
        "502": { description: Keycloak unreachable - refused before deleting anything (the tenant's logins could not be enumerated). }
  /api/v1/admin/tenants/{tenantId}/sites:
    get:
      tags: [admin]
      summary: List the sites of a tenant (platform-admin, cross-tenant).
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The tenant's sites.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
    post:
      tags: [admin]
      summary: Create a site for a tenant (platform-admin, cross-tenant).
      description: >
        Creates a site for the tenant named in the PATH, using the BYPASSRLS admin
        datasource. The tenant is validated to exist first. The tenant is never
        taken from client-controlled body data, so this is a deliberate,
        role-gated cross-tenant write - customer site access stays RLS-scoped.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteRequest" }
      responses:
        "201":
          description: Site created for the tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { description: Invalid input (blank name, out-of-range coordinates, or bad zone). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
  /api/v1/admin/tenants/{tenantId}/users:
    get:
      tags: [admin]
      summary: List the customer users (Portal-Users) of a tenant.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Users carrying this tenant's `tenant_id` in Keycloak.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
    post:
      tags: [admin]
      summary: Provision a customer user into a tenant.
      description: >
        Creates a Keycloak user with the `tenant_id` attribute and the customer
        (`operator`) realm role, so the account is tenant-scoped by the existing
        OIDC + RLS spine. An optional password is set (temporary for an invite
        flow). A duplicate username/email yields 409.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Customer user created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant not found. }
        "409": { description: A user with that username or email already exists. }
  /api/v1/admin/provisioned-devices:
    get:
      tags: [admin]
      summary: List the manufacturing registry of sticker Geräte-IDs.
      description: >
        Platform-wide view of every provisioned (manufactured) device ID and its
        claim state - which tenant, if any, has connected it.
      responses:
        "200":
          description: All provisioned devices.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProvisionedDevice" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Register a manufactured sticker Geräte-ID.
      description: >
        Registers a device ID (canonicalized like the claim path - trimmed,
        uppercased) so a customer can claim it. Only sticker-format IDs
        (prefix `VP-`) are accepted. Idempotent - re-registering an existing ID
        returns 200 with the unchanged entry.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProvisionDeviceRequest" }
      responses:
        "200":
          description: Already registered - the existing, unchanged entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvisionedDevice" }
        "201":
          description: Geräte-ID registered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProvisionedDevice" }
        "400": { description: Blank or non-sticker-format (missing VP- prefix) Geräte-ID. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/provisioned-devices/{externalRef}:
    delete:
      tags: [admin]
      summary: Remove a wrongly registered sticker Geräte-ID from the registry.
      description: >
        Refused (409) while a customer's claim references the ID - the device
        must be unclaimed first. The path ref is canonicalized like the claim
        path (trimmed, uppercased).
      parameters:
        - { name: externalRef, in: path, required: true, schema: { type: string }, example: VP-1234-ABCD }
      responses:
        "204": { description: Registry entry removed; the ID can no longer be claimed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: No such registry entry. }
        "409": { description: The ID is claimed by a customer; unclaim the device first. }
  /api/v1/admin/tenants/{tenantId}/users/{userId}:
    put:
      tags: [admin]
      summary: Update a customer user's profile (email/name).
      description: The username is the login identity and stays immutable.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { description: Invalid email or over-long name. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant or user not found (including a user outside this tenant). }
        "409": { description: A user with that email already exists. }
    delete:
      tags: [admin]
      summary: Permanently delete a customer user's login.
      description: >
        Admins can never delete their OWN account (409) - the guard fires
        before any tenant check.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: User deleted; their login stops working immediately. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant or user not found (including a user outside this tenant). }
        "409": { description: Attempt to delete one's own account. }
  /api/v1/admin/tenants/{tenantId}/users/{userId}/disable:
    post:
      tags: [admin]
      summary: Disable a customer user (blocks their logins).
      description: >
        Admins can never disable their OWN account (409) - the guard fires
        before any tenant check.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant or user not found (including a user outside this tenant). }
        "409": { description: Attempt to disable one's own account. }
  /api/v1/admin/tenants/{tenantId}/users/{userId}/enable:
    post:
      tags: [admin]
      summary: Re-enable a disabled customer user.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant or user not found (including a user outside this tenant). }
  /api/v1/admin/tenants/{tenantId}/users/{userId}/reset-password:
    post:
      tags: [admin]
      summary: Reset a customer user's password (support lever).
      description: >
        The platform has no SMTP and therefore no self-service password reset;
        this is how support recovers a customer who forgot their password or
        locked themselves out guessing. Sets the new password (`temporary`
        defaults to true: must be changed on next login) and lifts any
        brute-force lockout so the new password works immediately.
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "200":
          description: Password reset; any lockout lifted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { description: Invalid password (must be 8-128 characters). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Tenant or user not found (including a user outside this tenant). }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak-issued OIDC access token (realm `voltpilot`). Must carry a
        `tenant_id` claim; the API sets it as the Postgres `app.tenant_id`
        session variable to drive Row-Level-Security.

        Tenant switcher (Portal-Admins only): a token with the `platform-admin`
        realm role (which carries NO tenant_id) may select a tenant per request
        via the `X-Tenant-Id` header (UUID). The customer endpoints then serve
        that tenant's data through the same RLS scoping the customer gets. The
        header is IGNORED for every non-admin token, so it can never widen a
        customer's access; an admin without the header sees zero rows
        (default-deny).
  responses:
    Unauthorized:
      description: Missing, expired or invalid Bearer token.
    Forbidden:
      description: Authenticated, but lacking the required role (e.g. a customer calling an admin endpoint).
  schemas:
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        segment: { type: string, enum: [CI, B2C], example: CI }
        plan: { type: string, example: mvp }
        createdAt: { type: string, format: date-time }
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        segment: { type: string, enum: [CI, B2C], description: "Defaults to 'CI'." }
...
    UpdateTenantRequest:
      type: object
      required: [name]
      description: Same validation as CreateTenantRequest.
      properties:
        name: { type: string }
        segment: { type: string, enum: [CI, B2C], description: "Defaults to 'CI'." }
    DeleteTenantRequest:
      type: object
      required: [confirmName]
      properties:
        confirmName:
          type: string
          description: Must equal the tenant's exact name (type-to-confirm fat-finger protection).
    TenantOffboardingReport:
      type: object
      description: >
        What the offboarding removed. The database part is transactional
        (all-or-nothing); `failedUsers` lists Keycloak logins that could not be
        deleted and need manual cleanup.
      properties:
        tenantId: { type: string, format: uuid }
        tenantName: { type: string }
        deletedSites: { type: integer }
        deletedDevices: { type: integer }
        deletedTelemetryRows: { type: integer, format: int64 }
        deletedUsers:
          type: array
          items: { type: string }
        failedUsers:
          type: array
          items: { type: string }
...